(PC-9335): increase pool size (#322)
* (PC-9335): increase pool size

* (PC-9335): debug connection engine

* (PC-9335): mock connection for CI and update tests
