+++ recommendation/api/tests/recommendation_test.py
@@ -17,9 +17,7 @@ from recommendation import (
 @patch("recommendation.get_cold_start_scored_recommendations_for_user")
 @patch("recommendation.get_iris_from_coordinates")
 @patch("recommendation.save_recommendation")
-@patch("recommendation.create_db_connection")
 def test_get_final_recommendation_for_group_a_cold_start(
-    connection_mock: Mock,
     save_recommendation_mock: Mock,
     get_iris_from_coordinates_mock: Mock,
     get_cold_start_scored_recommendations_for_user_mock: Mock,
@@ -27,7 +25,6 @@ def test_get_final_recommendation_for_group_a_cold_start(
     app_config: Dict[str, Any],
 ):
     # Given
-    connection_mock.return_value = setup_database
     user_id = 113
     get_cold_start_scored_recommendations_for_user_mock.return_value = [
         {
@@ -49,7 +46,7 @@ def test_get_final_recommendation_for_group_a_cold_start(
 
     # When
     recommendations = get_final_recommendations(
-        user_id, 2.331289, 48.830719, app_config
+        user_id, 2.331289, 48.830719, app_config, setup_database
     )
 
     # Then
@@ -61,9 +58,7 @@ def test_get_final_recommendation_for_group_a_cold_start(
 @patch("recommendation.get_scored_recommendation_for_user")
 @patch("recommendation.get_iris_from_coordinates")
 @patch("recommendation.save_recommendation")
-@patch("recommendation.create_db_connection")
 def test_get_final_recommendation_for_group_a_algo(
-    connection_mock: Mock,
     save_recommendation_mock: Mock,
     get_iris_from_coordinates_mock: Mock,
     get_scored_recommendation_for_user_mock: Mock,
@@ -72,7 +67,6 @@ def test_get_final_recommendation_for_group_a_algo(
     app_config: Dict[str, Any],
 ):
     # Given
-    connection_mock.return_value = setup_database
     user_id = 111
     get_intermediate_recommendations_for_user_mock.return_value = [
         {
@@ -112,7 +106,7 @@ def test_get_final_recommendation_for_group_a_algo(
 
     # When
     recommendations = get_final_recommendations(
-        user_id, 2.331289, 48.830719, app_config
+        user_id, 2.331289, 48.830719, app_config, setup_database
     )
 
     # Then
@@ -125,9 +119,7 @@ def test_get_final_recommendation_for_group_a_algo(
 @patch("recommendation.get_iris_from_coordinates")
 @patch("recommendation.get_cold_start_types")
 @patch("recommendation.save_recommendation")
-@patch("recommendation.create_db_connection")
 def test_get_final_recommendation_for_group_b(
-    connection_mock: Mock,
     save_recommendation_mock: Mock,
     get_cold_start_types: Mock,
     get_iris_from_coordinates_mock: Mock,
@@ -137,7 +129,6 @@ def test_get_final_recommendation_for_group_b(
     app_config: Dict[str, Any],
 ):
     # Given
-    connection_mock.return_value = setup_database
     user_id = 112
     get_intermediate_recommendations_for_user_mock.return_value = [
         {
@@ -176,7 +167,7 @@ def test_get_final_recommendation_for_group_b(
     get_iris_from_coordinates_mock.return_value = 1
 
     recommendations = get_final_recommendations(
-        user_id, 2.331289, 48.830719, app_config
+        user_id, 2.331289, 48.830719, app_config, setup_database
     )
 
     # Then
@@ -192,9 +183,7 @@ def test_get_final_recommendation_for_group_b(
 @patch("recommendation.get_cold_start_types")
 @patch("recommendation.get_iris_from_coordinates")
 @patch("recommendation.save_recommendation")
-@patch("recommendation.create_db_connection")
 def test_get_final_recommendation_for_new_user(
-    connection_mock: Mock,
     save_recommendation_mock: Mock,
     get_iris_from_coordinates_mock: Mock,
     get_cold_start_types: Mock,
@@ -206,7 +195,6 @@ def test_get_final_recommendation_for_new_user(
     app_config: Dict[str, Any],
 ):
     # Given
-    connection_mock.return_value = setup_database
     user_id = 113
     get_cold_start_types.return_value = ["type2", "type3"]
     get_intermediate_recommendations_for_user.return_value = [
@@ -237,7 +225,7 @@ def test_get_final_recommendation_for_new_user(
 
     # When
     recommendations = get_final_recommendations(
-        user_id, 2.331289, 48.830719, app_config
+        user_id, 2.331289, 48.830719, app_config, setup_database
     )
     try:
         get_intermediate_recommendations_for_user.assert_called()
